+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/jwtpayload/KeyAttestationClaims.kt
@@ -1,0 +1,52 @@
+package at.asitplus.signum.indispensable.josef.jwtpayload
+
+import at.asitplus.signum.indispensable.josef.JsonWebKey
+import at.asitplus.signum.indispensable.josef.JwtClaimNames.IanaRegistered
+import at.asitplus.signum.indispensable.josef.JwtClaimNames.UnregisteredClaims
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+
+@Deprecated("Will move into VCK next release")
+@Serializable
+data class KeyAttestationClaims(
+    /**
+     * Array of attested keys from the same key storage component using the syntax of JWK as defined in RFC7517.
+     */
+    @SerialName(UnregisteredClaims.EudiTs3Claims.ATTESTED_KEYS)
+    val attestedKeys: Collection<JsonWebKey>,
+
+    /**
+     * Optional. Array of case-sensitive strings that assert the attack potential resistance of the key storage
+     * component and its keys attested in the attested_keys parameter. This specification defines initial values in
+     * Appendix D.2.
+     */
+    @SerialName(UnregisteredClaims.EudiTs3Claims.KEY_STORAGE)
+    val keyStorage: Collection<String>? = null,
+
+    /**
+     * Optional. Array of case-sensitive strings that assert the attack potential resistance of the user authentication
+     * methods allowed to access the private keys from the [attestedKeys] parameter.
+     * This specification defines initial values in Appendix D.2.
+     */
+    @SerialName(UnregisteredClaims.EudiTs3Claims.USER_AUTHENTICATION)
+    val userAuthentication: Collection<String>? = null,
+
+    /**
+     * Optional. A String that contains a URL that links to the certification of the key storage component.
+     * EUDI TS3 REQUIRED
+     */
+    @SerialName(UnregisteredClaims.EudiTs3Claims.CERTIFICATION)
+    val certification: String,
+
+    /**
+     * EUDI TS3 WUA 1.5: REQUIRED.
+     * status list reference for the attested key storage and the time until which the Wallet
+     * Provider commits to maintaining the referenced status.
+     */
+    @SerialName(UnregisteredClaims.EudiTs3Claims.KEY_STORAGE_STATUS)
+    val keyStorageStatus: KeyStorageStatus,
+
+    @SerialName(IanaRegistered.ClaimNames.OpenIdConnectCore.NONCE)
+    val nonce: String? = null,
+)